fastify auth server poc
