+++ apps/api/package.json
@@ -13,18 +13,21 @@
     "dev": "ts-node src/main.ts"
   },
   "devDependencies": {
-    "@types/fastify-cors": "",
+    "@types/bcrypt": "^5.0.0",
     "@types/jsonwebtoken": "^8.5.8",
     "@types/node": "^17.0.23",
+    "@types/passport-local": "^1.0.35",
     "ts-node": "^10.7.0",
     "typescript": "^4.6.3"
   },
   "dependencies": {
+    "@fastify/cookie": "^9.0.4",
+    "@fastify/cors": "^8.3.0",
+    "@fastify/session": "^10.4.0",
+    "bcrypt": "^5.0.1",
     "database": "*",
     "dotenv": "^16.0.0",
     "fastify": "4.22.2",
-    "fastify-cors": "",
-    "fastify-error": "3.3.0",
     "jsonwebtoken": "9.0.2"
   }
 }
