--- apps/client/next-swagger-doc.json
@@ -1,11 +1,0 @@
-{
-    "apiFolder": "pages/api",
-    "schemaFolders": ["models"],
-    "definition": {
-      "openapi": "3.0.0",
-      "info": {
-        "title": "Peppermint Swagger api docs",
-        "version": "1.0"
-      }
-    }
-  }
