+++ apps/client/next.config.js
@@ -18,25 +18,4 @@ module.exports = removeImports({
     return config
   },
   
-  // webpack: (config, options) => {
-
-  //   config.resolve.alias = {
-  //     ...config.resolve.alias,
-  //     'react-dom$': 'react-dom/profiling',
-  //     'scheduler/tracing': 'scheduler/tracing-profiling',
-  //   };
-
-  //   // Only disable function name mangling on the development enviroment
-  //   if (process.env.NODE_ENV === 'development') {
-  //       const terser = config.optimization.minimizer.find((plugin) => plugin instanceof TerserPlugin);
-  //       if (terser) {
-  //         terser.options.terserOptions = {
-  //           ...terser.options.terserOptions,
-  //           keep_classnames: true,
-  //           keep_fnames: true,
-  //         };
-  //       }
-  //   }
-  //   return config;
-  // }
 });
