+++ apps/client/package.json
@@ -40,6 +40,7 @@
     "antd": "^4.18.3",
     "axios": "^0.25.0",
     "bcrypt": "^5.0.1",
+    "cookies-next": "^3.0.0",
     "database": "*",
     "dompurify": "^2.4.0",
     "email-templates": "^8.0.8",
@@ -75,8 +76,9 @@
   },
   "devDependencies": {
     "@prisma/nextjs-monorepo-workaround-plugin": "^5.2.0",
+    "@types/next": "^9.0.0",
     "@types/node": "17.0.4",
-    "@types/react": "17.0.38",
+    "@types/react": "^18.2.21",
     "@types/swagger-ui-react": "^4.11.0",
     "autoprefixer": "^10.4.0",
     "postcss": "^8.4.5",
